integration: extract macros/complex_pk
There was a test case in test_prepared_statement() that actually had
nothing to do with prepared statements, but was testing the ser/de
macros. It was extracted to a separate file.
