+++ scylla/tests/integration/macros/mod.rs
@@ -1,1 +1,2 @@
+mod complex_pk;
 mod hygiene;
